+++ src/utils/typed_array.hpp
@@ -24,6 +24,7 @@ namespace node_gdal {
 // Uint16Array  2  16-bit unsigned integer  unsigned short
 // Int32Array   4  32-bit twos complement  signed integer  int
 // Uint32Array  4  32-bit unsigned integer  unsigned int
+// Float16Array 2  16-bit IEEE floating point number  GFloat16 (GDAL type)
 // Float32Array 4  32-bit IEEE floating point number  float
 // Float64Array 8  64-bit IEEE floating point number  double
 
@@ -40,7 +41,9 @@ namespace node_gdal {
 //   GDT_CInt32 = 9,
 //   GDT_CFloat32 = 10,
 //   GDT_CFloat64 = 11,
-//   GDT_TypeCount = 12
+//   GDT_TypeCount = 12,
+//   GDT_Float16 = 15,
+//   GDT_CFloat16 = 16
 // }
 
 namespace TypedArray {