+++ CHANGELOG.md
@@ -5,15 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-#
+# [3.11.0]
 
 ### Added
 
+ - GDAL 3.11.0
  - Add `libaec` and AEC encoding support for GRIB2 files - including those provided by Météo France
+ - Add the new `LIBERTIFF` driver
+ - Initial `Float16` support through the [`@petamoriken/float16`](https://www.npmjs.com/package/@petamoriken/float16) polyfill
 
 ### Changed
  - Fix the `gdal.info()` example
  - Fix [mmomtchev/node-gdal-async#204](https://github.com/mmomtchev/node-gdal-async/issues/204), define `HAVE_TIFF` for other drivers using built-in TIFF support
+
+### Removed
+ - Drop all obsolete in GDAL 3.11 drivers, including `SDTS`
 
 ## [3.10.2] 2025-03-09